+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## New Features
 - PR #361 Prototypes for cusort functions
 - PR #357 Pagerank cpp API
+- PR #366 Adds graph.degrees() function returning both in and out degree.
 
 
 ## Improvements
@@ -11,6 +12,7 @@
 - PR #365 Updates for setting device and vertex ids for snmg pagerank
 
 ## Bug Fixes
+- PR #368 Bump cudf dependency versions for cugraph conda packages
 - PR #354 Fixed bug in building a debug version
 - PR #360 Fixed bug in snmg coo2csr causing intermittent test failures.
 - PR #364 Fixed bug building or installing cugraph when conda isn't installed